Fixed downloader script by converting it to a BigIntegerField

+++ armaadmin/downloads/models.py
@@ -8,7 +8,7 @@ from django.utils.timezone import now
 class File(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     filename = models.CharField(max_length=255)
-    size = models.IntegerField()
+    size = models.BigIntegerField()
     full_path = models.CharField(max_length=255)
     relative_path = models.CharField(max_length=255)
     hash = models.CharField(max_length=8)
@@ -24,6 +24,6 @@ class FileSync(models.Model):
     deleted = models.IntegerField(blank=True, null=True)
     total = models.IntegerField(blank=True, null=True)
     previous_total = models.IntegerField(blank=True, null=True)
-    size = models.IntegerField(blank=True, null=True)
+    size = models.BigIntegerField(blank=True, null=True)
     started = models.DateTimeField(default=now, editable=False)
     finished = models.DateTimeField(editable=False, blank=True, null=True)
